1 (fix)

+++ contracts/LPTokenWrapper.sol
@@ -114,7 +114,7 @@ abstract contract LPTokenWrapper is Ownable {
         
 
         // Calculate fee reductions if applicable for users holding PRISM
-        uint256 userDiscount = 100.sub(userFeeReduceLvlDiscount);
+        uint256 userDiscount = uint256(100).sub(userFeeReduceLvlDiscount);
         uint256 feeReducedTax = tax.div(100).mul(userDiscount);
 
         // Transfer the tokens to user
